+++ api/src/main/java/ru/yajaneya/Spring2Geekbrains/api/core/ProductDto.java
@@ -2,6 +2,8 @@ package ru.yajaneya.Spring2Geekbrains.api.core;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 
+import java.math.BigDecimal;
+
 @Schema(description = "Модель продукта")
 public class ProductDto {
 
@@ -10,13 +12,13 @@ public class ProductDto {
     @Schema(description = "Название продукта", required = true, maxLength = 255, minLength = 3, example = "Коробка конфет")
     private String title;
     @Schema(description = "Цена продукта", required = true, example = "120")
-    private Integer price;
+    private BigDecimal price;
     private String category;
 
     public ProductDto() {
     }
 
-    public ProductDto(Long id, String title, Integer price, String category) {
+    public ProductDto(Long id, String title, BigDecimal price, String category) {
         this.id = id;
         this.title = title;
         this.price = price;
@@ -39,11 +41,11 @@ public class ProductDto {
         this.title = title;
     }
 
-    public Integer getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(Integer price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
